% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_run.R
\name{app_run}
\alias{app_run}
\title{Run the shiny application}
\usage{
app_run(...)
}
\arguments{
\item{...}{further arguments that can be passed to \code{\link[shiny]{shinyApp}} options}
}
\value{
shinyApp object, see also \code{\link[shiny]{shinyApp}}
}
\description{
Run the shiny application
}
