% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sidebar_link.R
\name{create_sidebar_link}
\alias{create_sidebar_link}
\title{Create sidebar link}
\usage{
create_sidebar_link(
  id,
  label,
  class = "link-body-emphasis d-inline-flex text-decoration-none rounded w-100",
  active = FALSE
)
}
\arguments{
\item{id}{input id for the link}

\item{label}{Label}

\item{class}{Bootstrap classes to apply to the link}

\item{active}{Whether this should be the active link}
}
\value{
tagList with a tags$li
}
\description{
Create sidebar link
}
