% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sidebar_menu_header.R
\name{create_sidebar_menu_header}
\alias{create_sidebar_menu_header}
\title{Create a sidebar menu header}
\usage{
create_sidebar_menu_header(
  title,
  title_class = "ps-2",
  data_bs_target,
  data_bs_toggle = "collapse",
  class =
    "btn btn-toggle d-inline-flex align-items-center rounded border-0 collapsed w-100",
  aria_expanded = "false"
)
}
\arguments{
\item{title}{Header title}

\item{title_class}{Bootstrap classes to apply to the title}

\item{data_bs_target}{'data-bs-target' attribute of the menu header}

\item{data_bs_toggle}{'data-bs-toggle' attribute of the menu header}

\item{class}{Bootstrap classes to apply to the menu header}

\item{aria_expanded}{'aria-expanded' attribute of the menu header. Whether
this sidebar menu is open or closed. Either "true" or "false".}
}
\value{
tagList with a tags$button
}
\description{
When clicked, it collapses it's contents (sidebar menus)
}
